+++ IdentityCore.sln
@@ -54,6 +54,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Identity.DefaultUI.WebSite", "test\WebSites\Identity.DefaultUI.WebSite\Identity.DefaultUI.WebSite.csproj", "{EA424B4D-0BE1-49AC-A106-CC6CC808A104}"
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.AspNetCore.Identity.SignIn", "src\Microsoft.AspNetCore.Identity.SignIn\Microsoft.AspNetCore.Identity.SignIn.csproj", "{C0E19D79-84F1-40A2-A8E9-5B5F4BAEA965}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.AspNetCore.Identity.SignIn.Cookies", "src\Microsoft.AspNetCore.Identity.SignIn.Cookies\Microsoft.AspNetCore.Identity.SignIn.Cookies.csproj", "{795AD298-C4E9-4CB0-ADEA-075D3EF12892}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -308,6 +310,22 @@ Global
 		{C0E19D79-84F1-40A2-A8E9-5B5F4BAEA965}.Release|x64.Build.0 = Release|Any CPU
 		{C0E19D79-84F1-40A2-A8E9-5B5F4BAEA965}.Release|x86.ActiveCfg = Release|Any CPU
 		{C0E19D79-84F1-40A2-A8E9-5B5F4BAEA965}.Release|x86.Build.0 = Release|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Debug|x64.Build.0 = Debug|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Debug|x86.Build.0 = Debug|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Release|Any CPU.Build.0 = Release|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Release|x64.ActiveCfg = Release|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Release|x64.Build.0 = Release|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Release|x86.ActiveCfg = Release|Any CPU
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -330,6 +348,7 @@ Global
 		{C47C1165-9F19-4DF8-ABA9-707ACEB3BDC7} = {52D59F18-62D2-4D17-8CF2-BE192445AF8E}
 		{EA424B4D-0BE1-49AC-A106-CC6CC808A104} = {C47C1165-9F19-4DF8-ABA9-707ACEB3BDC7}
 		{C0E19D79-84F1-40A2-A8E9-5B5F4BAEA965} = {0F647068-6602-4E24-B1DC-8ED91481A50A}
+		{795AD298-C4E9-4CB0-ADEA-075D3EF12892} = {0F647068-6602-4E24-B1DC-8ED91481A50A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {21D598B0-2383-4B22-826D-E7FB4921BD66}
