+++ src/Microsoft.AspNetCore.Identity.SignIn.Cookies/ISecurityStampValidator.cs
