+++ src/Microsoft.AspNetCore.Identity.SignIn.Cookies/ITwoFactorSecurityStampValidator.cs
