+++ src/Microsoft.AspNetCore.Identity.SignIn.Cookies/IdentityServiceCollectionExtensions.cs
@@ -1,0 +1,37 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authentication.Cookies;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    /// <summary>
+    /// Contains extension methods to <see cref="IServiceCollection"/> for configuring identity services.
+    /// </summary>
+    public static class IdentitySignInCookieServiceCollectionExtensions
+    {
+        /// <summary>
+        /// Configures the application cookie.
+        /// </summary>
+        /// <param name="services">The services available in the application.</param>
+        /// <param name="configure">An action to configure the <see cref="CookieAuthenticationOptions"/>.</param>
+        /// <returns>The services.</returns>
+        public static IServiceCollection ConfigureApplicationCookie(this IServiceCollection services, Action<CookieAuthenticationOptions> configure)
+            => services.Configure(IdentityConstants.ApplicationScheme, configure);
+
+        /// <summary>
+        /// Configure the external cookie.
+        /// </summary>
+        /// <param name="services">The services available in the application.</param>
+        /// <param name="configure">An action to configure the <see cref="CookieAuthenticationOptions"/>.</param>
+        /// <returns>The services.</returns>
+        public static IServiceCollection ConfigureExternalCookie(this IServiceCollection services, Action<CookieAuthenticationOptions> configure)
+            => services.Configure(IdentityConstants.ExternalScheme, configure);
+    }
+}
