+++ src/Microsoft.AspNetCore.Identity.SignIn.Cookies/Properties/debugSettings.json
@@ -1,0 +1,3 @@
+{
+  "Profiles": []
+}
