+++ src/Microsoft.AspNetCore.Identity.SignIn.Cookies/SecurityStampRefreshingPrincipalContext.cs
