+++ src/Microsoft.AspNetCore.Identity.SignIn.Cookies/SecurityStampValidator.cs
