+++ src/Microsoft.AspNetCore.Identity.SignIn.Cookies/SecurityStampValidatorOptions.cs
