+++ src/Microsoft.AspNetCore.Identity.SignIn.Cookies/TwoFactorSecurityStampValidator.cs
